Fixed the widget with the display of the number of exercises in the database after adding a new exercise or deleting it.
Added a tap gesture via an extension for all views, to close the keyboard when you tap anywhere on the screen except the keyboard itself.

The main view no longer slides down when the search keyboard appears.

The customization process with the Dark / Light mod has begun. (Changing colors depending on the mode.)

The Create Exercise button has been moved to a separate view file.

Added CreateProgramView, where you can create a new program.

Depending on the flag that determines whether the category to create an exercise is known or not, when you click different View