Unfortunately, due to my inattention, the new improvements which were made from March 16 to May 3 and were not saved to the github were deleted. During this period, I was on vacation from school and I did not study as intensively as before, which led to the loss of github settings in xCode, while fixing the project was rolled back to the time of March 16 by accident.
I wasn't able to restore the project until May 3, as a result more than a thousand if not more lines of code were deleted.

It was very sad to realize, but in any case the work will go on until the very end. And all the groundwork will have to be recreated from scratch, which will make some of the material better fixed in your head.

Do not despair and continue to work on the project
So in this commit the following was added :

----------------------------------------------------------------------------------------
Added animation and swipes algorithm in the calendar left to right with month change

Fixed display of selected date
Automatic date selection when turning on current day
Saving of selected date when switching month in the calendar and correct displaying it on return
Work on correctly minimizing calendar with correct displaying of row with selected day when calendar mode is minimized