Fixed a bug with the date of approaches.
Now approaches are correctly displayed for exercises in different workouts

Fixed duplication of programs in the database when adding in the application.

Fixed duplication of approaches

Fixed duplication of exercises

Added the ability to add exercises to existing programs with saving changes to the database, and developed the whole design for this purpose

Added display of already added exercises in the modified workout program with red highlighting and no possibility to add again.

Fixed some minor bugs affecting the application load and minor bugs

Translated with www.DeepL.com/Translator (free version)