Approaches are now saved in the database
Completely finished ( possible ) adding approaches with saving to the database

Correct display of approaches and their numbers in the main View of adding approaches

Added functionality to change the training program and add it as a workout with saving to the database.

Optimized code in some places.

All functionality moved to DataLoaderModel