Summary of work with the Realm Swift DataBase Library
Innovations:

Saving / Modifying / Deleting and Updating all possible data in the app ( Workouts / Exercises / Set in Exercise / Workout Programs )
Now automatically saved in the application and after exiting the application remain in memory, which allows full use of the application.

There may be some bugs in the future, as full testing has not been done, but at the moment no bugs have been noticed.

It will be polished better in the future.

The next functionality - Adding statistics in the form of graphs.