Version 1.0.3.
Fixed various bugs related to calendar optimization, because of which it was very laggy, jerky or reset when collapsing, deploying.

Changed design in some places (fonts, position, etc).

Fixed a bug when incorrectly deleting training from the database.
Fixed a bug when copying programs to the database.
Fixed an error with displaying training programs in the list of programs

Some fixed bugs are not described, because they are already forgotten in the process of fixing, but they are not significant.